Update README.md
"In this project, topology optimization of a structure was achieved by minimizing compliance, thereby increasing stiffness. Load conditions were applied to two structures: a truss and a building. Finite Element Method (FEM) techniques were used to form the stiffness matrix and displacement matrix. To perform the FEM analysis, meshes were generated, and each node was assigned degrees of freedom in the x and y directions. Subsequently, the Solid Isotropic Material with Penalization (SIMP) method was implemented for topology optimization in the computational design process. The primary goal was to reduce the volume fraction by 30% while ensuring the structure could still bear the applied loads.